hunter-crud/pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "huntercrud"
version = "0.1.0"
description = ""
authors = [ "Rakan Farhouda <rakanfarhouda@icloud.com>",]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2.16"
python-decouple = "^3.8"
coloredlogs = "^15.0.1"
colorama = "^0.4.6"
verboselogs = "^1.7"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
semver = "^3.0.2"
toml = "^0.10.2"
django-unfold = "^0.54.0"
pydantic = "^2.11.5"

[tool.poetry.group.docs.dependencies]
drf-spectacular = "^0.27.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
coverage = "^7.6.1"
pytest-cov = "^5.0.0"
pytest-profiling = "^1.7.0"
pytest-django = "^4.9.0"
pytest-html = "^4.1.1"
pytest-xdist = "^3.6.1"
faker = "^30.8.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
mypy = "^1.17.0"
django-coverage-plugin = "^3.1.1"
mypy-extensions = "^1.1.0"
django-stubs = "^5.2.2"
wemake-python-styleguide = "^1.3.0"
